% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_metabolism_gene_expression}
\alias{sim_metabolism_gene_expression}
\title{sim_metabolism_gene_expression function}
\usage{
sim_metabolism_gene_expression(
  module_gene_json_path,
  module_sim_flux_path,
  times = 900,
  precision = 1e-14
)
}
\arguments{
\item{sim_metabolism_gene_expression:}{Default:
module_gene_json_path
module_sim_flux_path
times=900
precision=0.00000000000001
metabolism_network_path}
}
\description{
Generate Simulated Metabolic Flux Data.
}
\examples{
sim_metabolism_gene_expression(module_gene_json_path, module_sim_flux_path, times=900, precision=0.00000000000001)


}
\keyword{sim_metabolism_gene_expression}
